pyproject.toml
[project]
name = "personal-assistant-backend"
version = "0.1.0"
description = "Personal Assistant Backend with Personality and HTN Orchestration"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.28.0",
    "httpx>=0.24.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "chromadb>=0.4.0",
    "redis>=5.0.0",
    "alembic>=1.14.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ASYNC"]
ignore = []

[tool.pyright]
include = ["src"]
typeCheckingMode = "strict"
pythonVersion = "3.11"
pythonPlatform = "Linux"
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownArgumentType = false
reportUnknownLambdaType = false
reportGeneralTypeIssues = true

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
